fix(ui): give recentPlaces a production consumer module for the knip gate
Knip's production scan rejects test-only exports; the recents derivation
moves to recent-places.ts, imported by the place picker, and the unit
test follows it.
